split tera.h to several roles #964
